requirements.txt
alembic==1.11.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.3
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Share==0.1.1
Flask-SQLAlchemy==3.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
